added a text box to inform the player that the game has been won or lost.
added to the GameBoard class a GameState enum to identify the state of the game

added to the GameBoard class the CheckBoard method to count the number of unrevealed cells and check if it equals the number of mines.

previously, the player would not be informed that the game was won or lost.
now, the appropriate text when a mine is revealed or all cells without a mine has been revealed.

next step,
1) move UI related code out of the GameBoard class.
2) find a less expensive method to check if a board has reached a win state (all cells without a mine have been revealed).